+++ ChangeLog.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+Added:
+
+* use `-fc, --force-compile` to force compilation of the script (issue #4)
 
 
 ## [0.4.0][0.4.0] - 27/02/2016